+++ src/unix/sys-std.cpp
@@ -1060,7 +1060,7 @@ attribute_hidden int Rstd_ReadConsole(const char *prompt, unsigned char *buf, in
 	if (rl_rest) {
 	    /* remaining line data (too long line) from a previous call */
 	    size_t r = strlen((char *)rl_rest + rl_rest_offset);
-	    if (r < len) {
+	    if (r < size_t(len)) {
 		memcpy(buf, rl_rest + rl_rest_offset, r);
 		buf[r] = '\0'; /* buf[r-1] is \n */
 		R_Free(rl_rest);
